% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{n_clusters}
\alias{n_clusters}
\title{Number of clusters found}
\usage{
n_clusters(x)
}
\arguments{
\item{x}{An `alethia_clusters` object.}
}
\value{
Integer count of distinct clusters.
}
\description{
Number of clusters found
}
\examples{
bag <- function(x) {
  t(sapply(
    strsplit(tolower(x), ""),
    function(ch) tabulate(match(ch, letters), 26)
  ))
}
n_clusters(cluster_entities(c("color", "colour", "zebra"), bag, floor = 0.8))
}
